move to log4j2 package
