+++ metadata-provider/src/main/resources/META-INF/services/org.apache.logging.log4j.core.util.ContextDataProvider
@@ -1,1 +1,1 @@
-com.dynatrace.logs.log4j.v2.DynatraceMetadataContextDataProvider
+DynatraceMetadataContextDataProvider
